Split to 5 separated packages
