--- app/Server/Main.hs
@@ -1,6 +1,0 @@
-module Main where
-
-import Cardano.Server.TestingServer.Main (runTestingServer)
-
-main :: IO ()
-main = runTestingServer
