+++ cabal.project
@@ -19,7 +19,12 @@ index-state:
   , hackage.haskell.org 2022-12-22T00:00:00Z
   , cardano-haskell-packages 2022-12-22T00:00:00Z
 
-packages: ./.
+packages:
+    cardano-server
+    cardano-server-client
+    cardano-server-example
+    cardano-server-client-example
+    cardano-server-test
 
 source-repository-package
   type: git
