+++ cardano-server-client-example/app/Main.hs
@@ -1,0 +1,6 @@
+module Main where
+
+import Cardano.Server.Client.Example.Main (runTestingClient)
+
+main :: IO ()
+main = runTestingClient
