+++ cardano-server-client-example/cardano-server-client-example.cabal
@@ -1,0 +1,51 @@
+cabal-version:      3.0
+
+name:               cardano-server-client-example
+version:            0.1.0.0
+
+author:             Grigory Petrov
+                    Vladimir Sinyakov
+maintainer:         team@encoins.io
+
+-- Plutus language settings
+common lang
+    default-language:   Haskell2010
+    ghc-options:
+      -fwarn-redundant-constraints
+      -Wall
+      -Widentities
+      -Wincomplete-uni-patterns
+      -Wincomplete-record-updates
+      -Wnoncanonical-monad-instances
+      -Wredundant-constraints
+      -Wno-unused-do-bind
+      
+      -- See Plutus Tx readme
+      -fno-ignore-interface-pragmas
+      -fno-omit-interface-pragmas
+      -fno-specialise
+      -fno-strictness
+      -fobject-code       
+      -fplugin-opt PlutusTx.Plugin:defer-errors
+
+library
+    import: lang
+    exposed-modules: 
+      Cardano.Server.Client.Example.Main
+    other-modules:
+    build-depends:
+      base                          >= 4.9 && < 5,
+      cardano-server-client                  -any,
+      cardano-server-example                 -any,
+      optparse-applicative                   -any,
+      plutus-tx                              -any,
+      random                                 -any,
+    hs-source-dirs: src
+
+executable cardano-server-client-example
+  default-language: Haskell2010
+  hs-source-dirs: app
+  main-is: Main.hs
+  build-depends: 
+    base                          >= 4.9 && < 5,
+    cardano-server-client-example              ,
