+++ cardano-server-client/cardano-server-client.cabal
@@ -1,0 +1,53 @@
+cabal-version:      3.0
+
+name:               cardano-server-client
+version:            0.1.0.0
+
+author:             Grigory Petrov
+                    Vladimir Sinyakov
+maintainer:         team@encoins.io
+
+-- Plutus language settings
+common lang
+    default-language:   Haskell2010
+    ghc-options:
+      -fwarn-redundant-constraints
+      -Wall
+      -Widentities
+      -Wincomplete-uni-patterns
+      -Wincomplete-record-updates
+      -Wnoncanonical-monad-instances
+      -Wredundant-constraints
+      -Wno-unused-do-bind
+      
+      -- See Plutus Tx readme
+      -fno-ignore-interface-pragmas
+      -fno-omit-interface-pragmas
+      -fno-specialise
+      -fno-strictness
+      -fobject-code       
+      -fplugin-opt PlutusTx.Plugin:defer-errors
+
+library
+    import: lang
+    exposed-modules: 
+      Cardano.Server.Client.Class
+      Cardano.Server.Client.Client
+      Cardano.Server.Client.Default
+      Cardano.Server.Client.Opts
+    other-modules:
+    default-extensions:
+      PartialTypeSignatures
+    build-depends:
+      base                          >= 4.9 && < 5,
+      cardano-server                         -any,
+      aeson                                  -any,
+      bytestring                             -any,
+      data-default                           -any,
+      http-client                            -any,
+      http-types                             -any,
+      mtl                                    -any,
+      optparse-applicative                   -any,
+      random                                 -any,
+      text                                   -any,
+    hs-source-dirs: src
