+++ cardano-server-example/app/Main.hs
@@ -1,0 +1,6 @@
+module Main where
+
+import Cardano.Server.Example.Main (runTestingServer)
+
+main :: IO ()
+main = runTestingServer
