+++ cardano-server-example/cardano-server-example.cabal
@@ -1,0 +1,59 @@
+cabal-version:      3.0
+
+name:               cardano-server-example
+version:            0.1.0.0
+
+author:             Grigory Petrov
+                    Vladimir Sinyakov
+maintainer:         team@encoins.io
+
+-- Plutus language settings
+common lang
+    default-language:   Haskell2010
+    ghc-options:
+      -fwarn-redundant-constraints
+      -Wall
+      -Widentities
+      -Wincomplete-uni-patterns
+      -Wincomplete-record-updates
+      -Wnoncanonical-monad-instances
+      -Wredundant-constraints
+      -Wno-unused-do-bind
+      
+      -- See Plutus Tx readme
+      -fno-ignore-interface-pragmas
+      -fno-omit-interface-pragmas
+      -fno-specialise
+      -fno-strictness
+      -fobject-code       
+      -fplugin-opt PlutusTx.Plugin:defer-errors
+
+library
+    import: lang
+    exposed-modules: 
+      Cardano.Server.Example.Main
+      Cardano.Server.Example.OnChain
+      Cardano.Server.Example.OffChain
+    other-modules:
+    build-depends:
+      base                          >= 4.9 && < 5,
+      cardano-server                         -any,
+      cardano-ledger-alonzo                  -any,
+      exceptions                             -any,
+      mtl                                    -any,
+      optparse-applicative                   -any,
+      plutus-apps-extra                      -any,
+      plutus-ledger                          -any,
+      plutus-ledger-api                      -any,
+      plutus-script-utils                    -any,
+      plutus-tx                              -any,
+      plutus-tx-plugin                       -any,
+    hs-source-dirs: src
+
+executable cardano-server-example
+  default-language: Haskell2010
+  hs-source-dirs: app
+  main-is: Main.hs
+  build-depends: 
+    base                          >= 4.9 && < 5,
+    cardano-server-example                     ,
