+++ cardano-server-test/cardano-server-test.cabal
@@ -1,0 +1,59 @@
+cabal-version:      3.0
+
+name:               cardano-server-test
+version:            0.1.0.0
+
+author:             Grigory Petrov
+                    Vladimir Sinyakov
+maintainer:         team@encoins.io
+
+-- Plutus language settings
+common lang
+    default-language:   Haskell2010
+    ghc-options:
+      -fwarn-redundant-constraints
+      -Wall
+      -Widentities
+      -Wincomplete-uni-patterns
+      -Wincomplete-record-updates
+      -Wnoncanonical-monad-instances
+      -Wredundant-constraints
+      -Wno-unused-do-bind
+      
+      -- See Plutus Tx readme
+      -fno-ignore-interface-pragmas
+      -fno-omit-interface-pragmas
+      -fno-specialise
+      -fno-strictness
+      -fobject-code       
+      -fplugin-opt PlutusTx.Plugin:defer-errors
+
+test-suite cardano-server-test
+  Type:                 exitcode-stdio-1.0
+  Default-Language:     Haskell2010
+  Hs-Source-Dirs:       test
+  Ghc-Options:          -rtsopts
+                        -fplugin-opt PlutusTx.Plugin:defer-errors
+  Main-Is:              Main.hs
+  Other-Modules:        DefaultClient
+                        Internal
+                        Reference
+                        TestingServer
+  Build-Depends:
+    base                          >= 4.9 && < 5,
+    cardano-server                         -any,
+    cardano-server-example                 -any,
+    cardano-server-client                  -any,
+    cardano-server-client-example          -any,
+    aeson                                  -any,
+    bytestring                             -any,
+    containers                             -any,
+    data-default                           -any,
+    directory                              -any,
+    mtl                                    -any,
+    plutus-apps-extra                      -any,
+    plutus-tx                              -any,
+    plutus-ledger                          -any,
+    plutus-ledger-constraints              -any,
+    text                                   -any,
+    time                                   -any
