+++ cardano-server-test/test/Main.hs
@@ -1,0 +1,7 @@
+module Main where
+
+import TestingServer (testFundsAllTS)
+import Reference     (runReferenceTest)
+
+main :: IO ()
+main = runReferenceTest
