--- cardano-server.cabal
@@ -1,155 +1,0 @@
-cabal-version:      3.0
-
-name:               cardano-server
-version:            0.1.0.0
-
-author:             Grigory Petrov
-                    Vladimir Sinyakov
-maintainer:         team@encoins.io
-
--- Plutus language settings
-common lang
-    default-language:   Haskell2010
-    ghc-options:
-      -fwarn-redundant-constraints
-      -Wall
-      -Widentities
-      -Wincomplete-uni-patterns
-      -Wincomplete-record-updates
-      -Wnoncanonical-monad-instances
-      -Wredundant-constraints
-      -Wno-unused-do-bind
-      
-      -- See Plutus Tx readme
-      -fno-ignore-interface-pragmas
-      -fno-omit-interface-pragmas
-      -fno-specialise
-      -fno-strictness
-      -fobject-code       
-      -fplugin-opt PlutusTx.Plugin:defer-errors
-
-library
-    import: lang
-    exposed-modules: 
-      Cardano.Server.Client.Class
-      Cardano.Server.Client.Client
-      Cardano.Server.Client.Default
-      Cardano.Server.Client.Opts
-      Cardano.Server.Class
-      Cardano.Server.Config
-      Cardano.Server.Endpoints.Funds 
-      Cardano.Server.Endpoints.Ping
-      Cardano.Server.Endpoints.Tx.Class
-      Cardano.Server.Endpoints.Tx.New
-      Cardano.Server.Endpoints.Tx.Server
-      Cardano.Server.Endpoints.Tx.Submit
-      Cardano.Server.Error
-      Cardano.Server.Input
-      Cardano.Server.Internal
-      Cardano.Server.Main
-      Cardano.Server.TestingServer.Main
-      Cardano.Server.TestingServer.OnChain
-      Cardano.Server.TestingServer.OffChain
-      Cardano.Server.Tx
-      Cardano.Server.Utils.Logger
-      Cardano.Server.Utils.Wait
-    other-modules:
-    default-extensions:
-      PartialTypeSignatures
-    build-depends:
-      base                          >= 4.9 && < 5,
-      aeson                                  -any,
-      aeson-casing                           -any,
-      bytestring                             -any,
-      cardano-addresses                      -any,
-      cardano-api                            -any,
-      cardano-crypto-class                   -any,
-      cardano-ledger-alonzo                  -any,
-      cardano-ledger-babbage                 -any,
-      cardano-ledger-byron                   -any,
-      cardano-ledger-core                    -any,
-      cardano-ledger-shelley                 -any,
-      cardano-node-emulator                  -any,
-      cardano-streaming                      -any,
-      cardano-wallet-core                    -any,
-      containers                             -any,
-      data-default                           -any,
-      directory                              -any,
-      exceptions                             -any,
-      extra                                  -any,
-      filepath                               -any,
-      hex-text                               -any,
-      http-client                            -any,
-      http-types                             -any,
-      mtl                                    -any,
-      optparse-applicative                   -any,
-      plutus-apps-extra                      -any,
-      plutus-chain-index                     -any,
-      plutus-chain-index-core                -any,
-      plutus-contract                        -any,
-      plutus-contract-certification          -any,
-      plutus-example                         -any,
-      plutus-ledger                          -any,
-      plutus-ledger-api                      -any,
-      plutus-ledger-constraints              -any,
-      plutus-pab                             -any,
-      plutus-script-utils                    -any,
-      plutus-tx                              -any,
-      plutus-tx-constraints                  -any,
-      plutus-tx-extra                        -any,
-      plutus-tx-plugin                       -any,
-      plutus-use-cases                       -any,
-      prettyprinter                          -any,
-      random                                 -any,
-      servant                                -any,
-      servant-checked-exceptions             -any,
-      servant-server                         -any,
-      text                                   -any,
-      time                                   -any,
-      wai                                    -any,
-      wai-cors                               -any,
-      warp                                   -any
-    hs-source-dirs: src
-
-test-suite cardano-server-test
-  Type:                 exitcode-stdio-1.0
-  Default-Language:     Haskell2010
-  Hs-Source-Dirs:       test
-  Ghc-Options:          -rtsopts
-                        -fplugin-opt PlutusTx.Plugin:defer-errors
-  Main-Is:              Main.hs
-  Other-Modules:        Tests.DefaultClient
-                        Tests.Internal
-                        Tests.Reference
-                        Tests.TestingServer
-  Build-Depends:
-    base                          >= 4.9 && < 5,
-    aeson                                  -any,
-    bytestring                             -any,
-    cardano-server                         -any,
-    containers                             -any,
-    data-default                           -any,
-    directory                              -any,
-    mtl                                    -any,
-    plutus-apps-extra                      -any,
-    plutus-tx                              -any,
-    plutus-ledger                          -any,
-    plutus-ledger-constraints              -any,
-    text                                   -any,
-    time                                   -any
-
-executable testingServer
-  default-language: Haskell2010
-  hs-source-dirs: app/Server
-  main-is: Main.hs
-  build-depends: 
-    base                          >= 4.9 && < 5,
-    cardano-server                             ,
-
-executable testingClient
-  default-language: Haskell2010
-  hs-source-dirs: app/Client
-  main-is: Main.hs
-  build-depends: 
-    base                          >= 4.9 && < 5,
-    cardano-server                             ,
