+++ cardano-server/cardano-server.cabal
@@ -1,0 +1,79 @@
+cabal-version:      3.0
+
+name:               cardano-server
+version:            0.1.0.0
+
+author:             Grigory Petrov
+                    Vladimir Sinyakov
+maintainer:         team@encoins.io
+
+-- Plutus language settings
+common lang
+    default-language:   Haskell2010
+    ghc-options:
+      -fwarn-redundant-constraints
+      -Wall
+      -Widentities
+      -Wincomplete-uni-patterns
+      -Wincomplete-record-updates
+      -Wnoncanonical-monad-instances
+      -Wredundant-constraints
+      -Wno-unused-do-bind
+      
+      -- See Plutus Tx readme
+      -fno-ignore-interface-pragmas
+      -fno-omit-interface-pragmas
+      -fno-specialise
+      -fno-strictness
+      -fobject-code       
+      -fplugin-opt PlutusTx.Plugin:defer-errors
+
+library
+    import: lang
+    exposed-modules: 
+      Cardano.Server.Class
+      Cardano.Server.Config
+      Cardano.Server.Endpoints.Funds 
+      Cardano.Server.Endpoints.Ping
+      Cardano.Server.Endpoints.Tx.Class
+      Cardano.Server.Endpoints.Tx.New
+      Cardano.Server.Endpoints.Tx.Server
+      Cardano.Server.Endpoints.Tx.Submit
+      Cardano.Server.Error
+      Cardano.Server.Input
+      Cardano.Server.Internal
+      Cardano.Server.Main
+      Cardano.Server.Tx
+      Cardano.Server.Utils.Logger
+      Cardano.Server.Utils.Wait
+    other-modules:
+    build-depends:
+      base                          >= 4.9 && < 5,
+      aeson                                  -any,
+      aeson-casing                           -any,
+      bytestring                             -any,
+      cardano-api                            -any,
+      cardano-node-emulator                  -any,
+      containers                             -any,
+      data-default                           -any,
+      directory                              -any,
+      exceptions                             -any,
+      extra                                  -any,
+      filepath                               -any,
+      hex-text                               -any,
+      http-types                             -any,
+      mtl                                    -any,
+      plutus-apps-extra                      -any,
+      plutus-ledger                          -any,
+      plutus-ledger-api                      -any,
+      plutus-tx                              -any,
+      prettyprinter                          -any,
+      servant                                -any,
+      servant-checked-exceptions             -any,
+      servant-server                         -any,
+      text                                   -any,
+      time                                   -any,
+      wai                                    -any,
+      wai-cors                               -any,
+      warp                                   -any
+    hs-source-dirs: src
