+++ cardano-server/logs/client.log
@@ -1,0 +1,132 @@
+
+Tue Nov 29 22:40:55 2022 
+Starting client...
+
+Tue Nov 29 22:40:55 2022 
+New tokens to send:
+
+["\240\157\189\154\240\168\158\177\243\131\158\128"]
+
+Tue Nov 29 22:41:39 2022 
+Starting client...
+
+Tue Nov 29 22:41:39 2022 
+New tokens to send:
+
+["asd","aasd"]
+
+Tue Nov 29 22:42:01 2022 
+Starting client...
+
+Tue Nov 29 22:42:01 2022 
+New tokens to send:
+
+["*token*","*token*"]
+
+Wed Nov 30 20:18:39 2022 
+Starting client...
+
+Wed Nov 30 20:18:39 2022 
+New tokens to send:
+
+["asd"]
+
+Wed Nov 30 20:18:39 2022 
+Received response:
+Response {responseStatus = Status {statusCode = 500, statusMessage = "Internal Server Error"}, responseVersion = HTTP/1.1, responseHeaders = [("Transfer-Encoding","chunked"),("Date","Wed, 30 Nov 2022 17:18:39 GMT"),("Server","Warp/3.3.20"),("Content-Type","text/plain; charset=utf-8")], responseBody = "Something went wrong", responseCookieJar = CJ {expose = []}, responseClose' = ResponseClose, responseOriginalRequest = Request {
+  host                 = "localhost"
+  port                 = 3000
+  secure               = False
+  requestHeaders       = [("Content-Type","application/json")]
+  path                 = "/relayRequestMint"
+  queryString          = ""
+  method               = "POST"
+  proxy                = Nothing
+  rawBody              = False
+  redirectCount        = 10
+  responseTimeout      = ResponseTimeoutMicro 180000000
+  requestVersion       = HTTP/1.1
+  proxySecureMode      = ProxySecureWithConnect
+}
+}
+
+Wed Nov 30 20:36:26 2022 
+Starting client...
+
+Wed Nov 30 20:36:26 2022 
+New tokens to send:
+
+["asd"]
+
+Wed Nov 30 20:36:26 2022 
+Received response:
+Response {responseStatus = Status {statusCode = 500, statusMessage = "Internal Server Error"}, responseVersion = HTTP/1.1, responseHeaders = [("Transfer-Encoding","chunked"),("Date","Wed, 30 Nov 2022 17:36:26 GMT"),("Server","Warp/3.3.20"),("Content-Type","text/plain; charset=utf-8")], responseBody = "Something went wrong", responseCookieJar = CJ {expose = []}, responseClose' = ResponseClose, responseOriginalRequest = Request {
+  host                 = "localhost"
+  port                 = 3000
+  secure               = False
+  requestHeaders       = [("Content-Type","application/json")]
+  path                 = "/relayRequestMint"
+  queryString          = ""
+  method               = "POST"
+  proxy                = Nothing
+  rawBody              = False
+  redirectCount        = 10
+  responseTimeout      = ResponseTimeoutMicro 180000000
+  requestVersion       = HTTP/1.1
+  proxySecureMode      = ProxySecureWithConnect
+}
+}
+
+Wed Nov 30 20:38:26 2022 
+Starting client...
+
+Wed Nov 30 20:38:26 2022 
+New tokens to send:
+
+["asd"]
+
+Wed Nov 30 20:38:26 2022 
+Received response:
+Response {responseStatus = Status {statusCode = 500, statusMessage = "Internal Server Error"}, responseVersion = HTTP/1.1, responseHeaders = [("Transfer-Encoding","chunked"),("Date","Wed, 30 Nov 2022 17:38:26 GMT"),("Server","Warp/3.3.20"),("Content-Type","text/plain; charset=utf-8")], responseBody = "Something went wrong", responseCookieJar = CJ {expose = []}, responseClose' = ResponseClose, responseOriginalRequest = Request {
+  host                 = "localhost"
+  port                 = 3000
+  secure               = False
+  requestHeaders       = [("Content-Type","application/json")]
+  path                 = "/relayRequestMint"
+  queryString          = ""
+  method               = "POST"
+  proxy                = Nothing
+  rawBody              = False
+  redirectCount        = 10
+  responseTimeout      = ResponseTimeoutMicro 180000000
+  requestVersion       = HTTP/1.1
+  proxySecureMode      = ProxySecureWithConnect
+}
+}
+
+Wed Nov 30 20:38:39 2022 
+Starting client...
+
+Wed Nov 30 20:38:39 2022 
+New tokens to send:
+
+["asd"]
+
+Wed Nov 30 20:38:39 2022 
+Received response:
+Response {responseStatus = Status {statusCode = 204, statusMessage = "No Content"}, responseVersion = HTTP/1.1, responseHeaders = [("Date","Wed, 30 Nov 2022 17:38:39 GMT"),("Server","Warp/3.3.20"),("Content-Type","application/json;charset=utf-8")], responseBody = "", responseCookieJar = CJ {expose = []}, responseClose' = ResponseClose, responseOriginalRequest = Request {
+  host                 = "localhost"
+  port                 = 3000
+  secure               = False
+  requestHeaders       = [("Content-Type","application/json")]
+  path                 = "/relayRequestMint"
+  queryString          = ""
+  method               = "POST"
+  proxy                = Nothing
+  rawBody              = False
+  redirectCount        = 10
+  responseTimeout      = ResponseTimeoutMicro 180000000
+  requestVersion       = HTTP/1.1
+  proxySecureMode      = ProxySecureWithConnect
+}
+}
