--- test/Main.hs
@@ -1,7 +1,0 @@
-module Main where
-
-import Tests.TestingServer (testFundsAllTS)
-import Tests.Reference     (runReferenceTest)
-
-main :: IO ()
-main = runReferenceTest
